refactored api get chunk
